fix: modularize ui components (phases 1-2)
