+++ apps/tauri/src/renderer/routes/analyze.tsx
@@ -2,12 +2,9 @@ import {
   AlertCircle,
   AlertTriangle,
   Briefcase,
-  CheckCircle2,
-  ChevronDown,
   RefreshCw,
   ScanSearch,
   Sparkles,
-  Upload,
   Zap,
 } from 'lucide-react';
 import { AnimatePresence, motion } from 'motion/react';
@@ -15,16 +12,16 @@ import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { createFileRoute } from '@tanstack/react-router';
 
 import type { DocumentRecord } from '@ajh/shared';
-import { Button, TextArea } from '@ajh/ui';
+import { Button } from '@ajh/ui';
 
 import { PageTransition } from '@/components/layout/PageTransition';
 import { ModelSelector, useCanUseAI, useSelectedModel } from '@/components/ui/ModelSelector';
-import { ThinkingBubble } from '@/features/ai-generate/components/ThinkingBubble';
 import { ResumeInputCard } from '@/features/ai-workspace/components/ResumeInputCard';
 import { AnalysisATSRisks } from '@/features/analyze/components/AnalysisATSRisks';
 import { AnalysisLanguageMismatch } from '@/features/analyze/components/AnalysisLanguageMismatch';
 import { AnalysisLanguageRecommendations } from '@/features/analyze/components/AnalysisLanguageRecommendations';
 import { AnalysisMissingSkills } from '@/features/analyze/components/AnalysisMissingSkills';
+import { AnalysisProgress } from '@/features/analyze/components/AnalysisProgress';
 import { AnalysisRecommendations } from '@/features/analyze/components/AnalysisRecommendations';
 import { AnalysisRewrites } from '@/features/analyze/components/AnalysisRewrites';
 import { AnalysisScores } from '@/features/analyze/components/AnalysisScores';
@@ -32,9 +29,9 @@ import { AnalysisSectionAnalysis } from '@/features/analyze/components/AnalysisS
 import { AnalysisSkills } from '@/features/analyze/components/AnalysisSkills';
 import { AnalysisStrengths } from '@/features/analyze/components/AnalysisStrengths';
 import { AnalysisVerdict } from '@/features/analyze/components/AnalysisVerdict';
+import { CollapsibleInput } from '@/features/analyze/components/CollapsibleInput';
 import { cn } from '@/lib/cn';
 import { useTranslation } from '@/lib/i18n';
-import { transition } from '@/lib/motion';
 import { type AnalysisResult, runAnalysis } from '@/lib/resume-ai';
 import { useDocuments, useExtractText } from '@/services';
 import type { PromptQuality } from '@/store/preferences-schema';
@@ -44,7 +41,6 @@ import { useSessionStore } from '@/store/session-store';
 export const Route = createFileRoute('/analyze')({ component: Analyze });
 
 const ACCEPTED_EXTS = ['pdf', 'docx', 'txt', 'md', 'markdown'] as const;
-const ACCEPT_ATTR = '.pdf,.docx,.txt,.md,.markdown';
 const MAX_BYTES = 25 * 1024 * 1024;
 
 type Stage = 'idle' | 'running' | 'done';
@@ -403,282 +399,6 @@ function Analyze() {
         </div>
       </div>
     </PageTransition>
-  );
-}
-
-// ─── Collapsible input card (mirrors ai-generate FileInput) ──────────────────
-
-function CollapsibleInput({
-  label,
-  icon: Icon,
-  value,
-  onChange,
-  uploading,
-  onUpload,
-  placeholder,
-  disabled,
-  t,
-}: {
-  label: string;
-  icon: React.ElementType;
-  value: string;
-  onChange: (v: string) => void;
-  uploading: boolean;
-  onUpload: (f: File) => void;
-  placeholder: string;
-  disabled?: boolean;
-  t: (key: string) => string;
-}) {
-  const ref = useRef<HTMLInputElement>(null);
-  const [expanded, setExpanded] = useState(true);
-
-  return (
-    <div
-      className={cn(
-        'glass-graphite glass-highlight rounded-xl overflow-hidden transition-colors',
-        value ? 'border-brand/20' : ''
-      )}
-    >
-      <div className="flex items-center justify-between px-3 py-2.5">
-        <div className="flex items-center gap-2">
-          <Icon size={13} className={value ? 'text-brand-soft' : 'text-foreground/30'} />
-          <span className="text-xs font-medium text-foreground/70">{label}</span>
-          {value && <CheckCircle2 size={11} className="text-emerald-400" />}
-        </div>
-        <div className="flex items-center gap-2">
-          {!disabled && (
-            <>
-              <input
-                ref={ref}
-                type="file"
-                accept={ACCEPT_ATTR}
-                className="hidden"
-                onChange={(e) => {
-                  const f = e.target.files?.[0];
-                  if (f) onUpload(f);
-                  e.target.value = '';
-                }}
-              />
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => ref.current?.click()}
-                disabled={uploading || disabled}
-                className="flex items-center gap-1 rounded-md bg-white/[0.04] px-2 py-1 text-[10px] text-foreground/50 hover:text-foreground/80 transition-colors h-auto"
-              >
-                <Upload size={10} className={uploading ? 'animate-pulse' : ''} />
-                {uploading ? '…' : t('analyze.uploadButton')}
-              </Button>
-            </>
-          )}
-          <Button
-            variant="ghost"
-            size="sm"
-            onClick={() => setExpanded((e) => !e)}
-            className="text-foreground/30 hover:text-foreground/60 transition-colors h-auto p-1"
-          >
-            <ChevronDown
-              size={13}
-              className={cn('transition-transform', expanded && 'rotate-180')}
-            />
-          </Button>
-        </div>
-      </div>
-      <AnimatePresence initial={false}>
-        {expanded && (
-          <motion.div
-            initial={{ height: 0 }}
-            animate={{ height: 'auto' }}
-            exit={{ height: 0 }}
-            transition={transition.normal}
-            className="overflow-hidden"
-          >
-            <TextArea
-              value={value}
-              onChange={(e) => onChange(e.target.value)}
-              disabled={disabled}
-              placeholder={placeholder}
-              className="w-full bg-transparent px-3 py-2.5 text-xs text-foreground/75 placeholder:text-foreground/35 font-mono leading-relaxed disabled:opacity-40"
-              style={{ height: 140 }}
-              spellCheck={false}
-            />
-          </motion.div>
-        )}
-      </AnimatePresence>
-    </div>
-  );
-}
-
-// ─── Analysis progress ────────────────────────────────────────────────────────
-
-// Estimated total duration in ms — progress bar reaches 90% at this point then waits
-const ESTIMATED_MS = 50_000;
-
-function AnalysisProgress({
-  running,
-  stream,
-  thinking,
-  modelLoading,
-  tokenCount,
-  tokenStartMs,
-  t,
-}: {
-  running: boolean;
-  stream: string;
-  thinking?: string;
-  modelLoading?: boolean;
-  tokenCount?: number;
-  tokenStartMs?: number | null;
-  t: (key: string) => string;
-}) {
-  const PROGRESS_MESSAGES = [
-    t('analyze.progress.reading'),
-    t('analyze.progress.scanning'),
-    t('analyze.progress.calculating'),
-    t('analyze.progress.checking'),
-    t('analyze.progress.measuring'),
-    t('analyze.progress.identifying'),
-    t('analyze.progress.prioritising'),
-    t('analyze.progress.scoring'),
-    t('analyze.progress.generating'),
-    t('analyze.progress.risks'),
-    t('analyze.progress.writing'),
-    t('analyze.progress.finalising'),
-  ];
-
-  const [progress, setProgress] = useState(0); // 0–100
-  const [elapsed, setElapsed] = useState(0); // seconds
-  const [msgIdx, setMsgIdx] = useState(0);
-  const startRef = useRef(Date.now());
-
-  useEffect(() => {
-    if (!running) return;
-
-    startRef.current = Date.now();
-    setProgress(0);
-    setElapsed(0);
-    setMsgIdx(0);
-
-    const tick = setInterval(() => {
-      const ms = Date.now() - startRef.current;
-      setElapsed(Math.floor(ms / 1000));
-      // Ease toward 90% over ESTIMATED_MS, then crawl slowly after
-      const t = Math.min(ms / ESTIMATED_MS, 1);
-      const eased = t < 1 ? 90 * (1 - Math.pow(1 - t, 3)) : 90 + (ms - ESTIMATED_MS) / 3000;
-      setProgress(Math.min(eased, 99));
-    }, 400);
-
-    const msgTick = setInterval(() => {
-      setMsgIdx((i) => (i + 1) % PROGRESS_MESSAGES.length);
-    }, 3500);
-
-    return () => {
-      clearInterval(tick);
-      clearInterval(msgTick);
-    };
-  }, [running, PROGRESS_MESSAGES.length]);
-
-  const mins = Math.floor(elapsed / 60);
-  const secs = elapsed % 60;
-  const timer = mins > 0 ? `${mins}m ${secs}s` : `${secs}s`;
-  const eta =
-    elapsed > 5 && progress < 90
-      ? `~${Math.max(1, Math.round(ESTIMATED_MS / 1000 - elapsed))}s left`
-      : progress >= 90
-        ? t('analyze.progress.almostDone')
-        : '';
-
-  return (
-    <div className="mt-4 rounded-xl border border-white/[0.07] bg-white/[0.02] px-6 py-6 space-y-5">
-      {/* Top row */}
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-2">
-          <span className="h-1.5 w-1.5 animate-pulse rounded-full bg-brand" />
-          <span className="text-[10px] font-semibold uppercase tracking-[0.18em] text-foreground/35">
-            {t('analyze.running')}
-          </span>
-        </div>
-        <div className="flex items-center gap-3 text-[10px] text-foreground/30">
-          {eta && <span className="text-brand-soft/70">{eta}</span>}
-          <span>{timer}</span>
-        </div>
-      </div>
-
-      {/* Rotating message */}
-      <div className="relative h-6 overflow-hidden">
-        <AnimatePresence mode="wait">
-          <motion.p
-            key={msgIdx}
-            initial={{ opacity: 0, y: 10 }}
-            animate={{ opacity: 1, y: 0 }}
-            exit={{ opacity: 0, y: -10 }}
-            transition={transition.slow}
-            className="absolute inset-0 flex items-center text-sm text-foreground/60"
-          >
-            {PROGRESS_MESSAGES[msgIdx]}
-          </motion.p>
-        </AnimatePresence>
-      </div>
-
-      {/* Progress bar */}
-      <div className="space-y-1.5">
-        <div className="h-1.5 overflow-hidden rounded-full bg-white/[0.06]">
-          <motion.div
-            className="h-full rounded-full bg-gradient-to-r from-violet-700 via-brand to-brand-soft"
-            animate={{ width: `${progress}%` }}
-            transition={transition.progress}
-          />
-        </div>
-        <div className="flex justify-end text-[10px] text-foreground/20">
-          <span>{Math.round(progress)}%</span>
-        </div>
-      </div>
-
-      {/* Thinking output */}
-      {thinking && <ThinkingBubble thinking={thinking} done={stream.length > 0} />}
-
-      {/* Model loading indicator */}
-      {modelLoading && (
-        <div className="flex items-center gap-2 text-[10px] text-foreground/40">
-          <span className="h-1.5 w-1.5 animate-spin rounded-full border border-brand border-t-transparent" />
-          Loading model into memory...
-        </div>
-      )}
-
-      {/* Live stream output */}
-      <div className="rounded-lg border border-white/[0.05] bg-black/20 px-4 py-3 h-28 overflow-hidden relative">
-        {stream ? (
-          <pre className="font-mono text-[10px] leading-relaxed text-foreground/30 whitespace-pre-wrap break-all">
-            {stream.slice(-800)}
-          </pre>
-        ) : (
-          <div className="space-y-2 pt-1">
-            {[1, 0.7, 0.85, 0.5].map((w, i) => (
-              <div
-                key={i}
-                className="h-2 rounded-full bg-white/[0.05] animate-pulse"
-                style={{ width: `${w * 100}%`, animationDelay: `${i * 150}ms` }}
-              />
-            ))}
-          </div>
-        )}
-        {/* Fade out top so it looks like it's scrolling in from below */}
-        <div className="absolute inset-x-0 top-0 h-8 bg-gradient-to-b from-black/20 to-transparent pointer-events-none" />
-      </div>
-
-      {/* Token throughput */}
-      {tokenCount != null &&
-        tokenCount > 0 &&
-        (() => {
-          const elapsed = tokenStartMs ? (Date.now() - tokenStartMs) / 1000 : 0;
-          const tokPerSec = elapsed > 2 ? Math.round(tokenCount / elapsed) : null;
-          return (
-            <div className="text-[10px] text-foreground/25 text-right">
-              {tokenCount.toLocaleString()} tokens{tokPerSec ? ` · ~${tokPerSec} tok/s` : ''}
-            </div>
-          );
-        })()}
-    </div>
   );
 }
